Revert "misc: Merge SM-G900F kernel source code"
This reverts commit 67f19bfe81079abb9ca7f5d68f1990826e10576e.

+++ drivers/media/tdmb/fc8080/fc8080_spi.c
@@ -150,7 +150,7 @@ static s32 spi_bulkwrite(HANDLE handle, u16 addr, u8 command, u8 *data,
 	return BBM_OK;
 }
 
-static s32 spi_dataread(HANDLE handle, u16 addr, u8 command, u8 *data,
+static s32 spi_dataread(HANDLE handle, u8 addr, u8 command, u8 *data,
 			u32 length)
 {
 	s32 res = BBM_OK;
